requirements.txt
aniso8601==9.0.1
APScheduler==3.7.0
astroid==2.6.6
attrs==21.2.0
backports.entry-points-selectable==1.1.0
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
coverage==5.5
distlib==0.3.2
filelock==3.0.12
flake8==3.9.2
Flask==2.0.1
Flask-APScheduler==1.12.2
Flask-RESTful==0.3.9
idna==3.2
iniconfig==1.1.1
isort==5.9.3
itsdangerous==2.0.1
Jinja2==3.0.1
kafka-python==2.0.2
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
mccabe==0.6.1
packaging==21.0
platformdirs==2.2.0
pluggy==0.13.1
psycopg2-binary==2.9.1
py==1.10.0
pycodestyle==2.7.0
pyflakes==2.3.1
pylint==2.9.6
pyparsing==2.4.7
pytest==6.2.4
pytest-cov==2.12.1
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
responses==0.13.4
six==1.16.0
toml==0.10.2
tox==3.24.1
tzlocal==2.1
urllib3==1.26.6
virtualenv==20.7.1
Werkzeug==2.0.1
wrapt==1.12.1
